+++ CHANGES.md
@@ -19,6 +19,7 @@
     in favor of dedicated scopesCompile, scopesRun and scopesTest keys along with scoping information directly inside externalClasspath
 - Introduced aggregate scope to implement aggregate project dependency more cleanly
 - Removed `running`, `compiling`, `assembling` and `archiving` configurations, they are not necessary anymore
+- Removed `retrievingSources` and `retrievingDocs` configurations, use new keys `externalSources` and `externalDocs` instead
 
 # 0.13 2020-06-23
 - Add utility functions for source file and classpath generation, see the `wemi.generation` package